pyproject.toml
[tool.poetry]
name = "feinstaub-api"
version = "0.0.1"
description = "Api to save data from sensors (especially particulates sensors)."
repository = "https://github.com/opendata-stuttgart/feinstaub-api"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
authors = [
  "Opendata Stuttgart",
  "André Honeiser <andre@honeiser.de>"
]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
gunicorn = "^20.0.4"
django-dbbackup = "^3.3.0"
django-extensions = "^3.0.9"
django-filter = "^2.4.0"
django-redis = "^4.12.1"
djangorestframework = "^3.12.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
